+++ dashboard-ui/src/features/Header/Selectors/types.ts
@@ -2,3 +2,9 @@
  * Copyright 2025 Lincoln Institute of Land Policy
  * SPDX-License-Identifier: MIT
  */
+
+import { ComboboxItem } from '@mantine/core';
+
+export type Option = ComboboxItem & {
+    id: number;
+};
