+++ dashboard-ui/src/features/Map/config.tsx
@@ -4,7 +4,6 @@
  */
 
 import {
-    BasemapId,
     CustomListenerFunction,
     LayerType,
     MainLayerDefinition,
@@ -13,88 +12,16 @@ import {
 } from '@/components/Map/types';
 import {
     DataDrivenPropertyValueSpecification,
-    ExpressionSpecification,
     LayerSpecification,
     Map,
     Popup,
 } from 'mapbox-gl';
-import { basemaps } from '@/components/Map/consts';
-
-export const MAP_ID = 'main';
-
-export const BASEMAP = basemaps[BasemapId.Dark];
-
-export enum SourceId {
-    Regions = 'regions-source',
-    Basins = 'hu04',
-    Reservoirs = 'reservoirs-source',
-    SnowWater = 'snow-water',
-}
-
-export enum LayerId {
-    Regions = 'regions-main',
-    Basins = 'basins-main',
-    Reservoirs = 'reservoirs',
-    SnowWater = 'snow-water',
-}
-
-export enum SubLayerId {
-    RegionsBoundary = 'regions-boundary',
-    RegionsFill = 'regions-fill',
-    BasinsBoundary = 'basins-boundary',
-    BasinsFill = 'basins-fill',
-}
-
-export const allLayerIds = [
-    ...Object.values(LayerId),
-    ...Object.values(SubLayerId),
-];
-
-const TeacupStepExpression: ExpressionSpecification = [
-    'step',
-    ['var', 'storage'],
-    'default', // Below first step value
-    0.05,
-    'teacup-5',
-    0.1,
-    'teacup-10',
-    0.15,
-    'teacup-15',
-    0.2,
-    'teacup-20',
-    0.25,
-    'teacup-25',
-    0.3,
-    'teacup-30',
-    0.35,
-    'teacup-35',
-    0.4,
-    'teacup-40',
-    0.45,
-    'teacup-45',
-    0.5,
-    'teacup-50',
-    0.55,
-    'teacup-55',
-    0.6,
-    'teacup-60',
-    0.65,
-    'teacup-65',
-    0.7,
-    'teacup-70',
-    0.75,
-    'teacup-75',
-    0.8,
-    'teacup-80',
-    0.85,
-    'teacup-85',
-    0.9,
-    'teacup-90',
-    0.95,
-    'teacup-95',
-    1.0,
-    'teacup-100',
-];
+import {
+    SubLayerId,
+    LayerId,
+    SourceId,
+    ReserviorIconImageExpression,
+} from '@/features/Map/consts';
 
 /**********************************************************************
  * Define the various datasources this map will use
@@ -273,51 +200,7 @@ export const getLayerConfig = (
                 type: LayerType.Symbol,
                 source: SourceId.Reservoirs,
                 layout: {
-                    // 'icon-image': 'default',
-                    'icon-image': [
-                        'let',
-                        'capacity', // Variable name
-                        ['coalesce', ['get', 'Active Capacity'], 1], // Variable value
-                        'storage', // Variable name
-                        [
-                            '/',
-                            ['/', ['coalesce', ['get', 'Live Capcity'], 1], 2],
-                            ['coalesce', ['get', 'Active Capacity'], 1],
-                        ], // Variable value
-                        [
-                            'step',
-                            ['zoom'],
-                            TeacupStepExpression,
-
-                            ...[
-                                [0, 2010000],
-                                // [2, 985000],
-                                // [3, 745000],
-                                [4, 465000],
-                                [5, 320000],
-                                // [6, 250000],
-                                [7, 65000],
-                                [8, -1],
-                                // [9, 90000],
-                                // [10, 65000],
-
-                                // [12, 30000],
-                                // [13, 25000],
-                                // [14, 13000],
-                                // [15, 7000],
-                                // [16, 3300],
-                            ].flatMap(([zoom, capacity]) => [
-                                zoom, // At this zoom
-                                [
-                                    // Evaluate this expression
-                                    'case',
-                                    ['>=', ['var', 'capacity'], capacity],
-                                    TeacupStepExpression, // If GTE, evaluate sub-step expression
-                                    'default', // Fallback to basic point symbol
-                                ],
-                            ]),
-                        ],
-                    ],
+                    'icon-image': ReserviorIconImageExpression,
                     'icon-size': [
                         'let',
                         'capacity',
@@ -326,44 +209,18 @@ export const getLayerConfig = (
                             'step',
                             ['var', 'capacity'],
                             0.3,
-                            // 3300,
-                            // 0.31,
-                            // 7000,
-                            // 0.32,
-                            // 13000,
-                            // 0.34,
-                            // 25000,
-                            // 0.36,
-                            // 30000,
-                            // 0.38,
                             45000,
                             0.4,
-                            // 65000,
-                            // 0.42,
-                            // 90000,
-                            // 0.44,
-                            // 190000,
-                            // 0.46,
-                            // 250000,
-                            // 0.48,
                             320000,
                             0.5,
-                            // 465000,
-                            // 0.52,
-                            // 745000,
-                            // 0.54,
-                            // 985000,
-                            // 0.56,
-                            // 1500000,
-                            // 0.58,
                             2010000,
                             0.6,
                         ],
                     ],
 
                     'symbol-sort-key': [
-                        '*',
-                        ['coalesce', ['get', 'Active Capacity'], 1],
+                        'coalesce',
+                        ['get', 'Active Capacity'],
                         1,
                     ],
                     'icon-offset': [
