+++ dashboard-ui/src/providers/Map.tsx
@@ -5,7 +5,7 @@
 
 import { PropsWithChildren } from 'react';
 import { MapProvider } from '@/contexts/MapContexts';
-import { MAP_ID } from '@/features/Map/config';
+import { MAP_ID } from '@/features/Map/consts';
 
 /**
  * Provides Map Context to allow accessing maps across application
